Fixed step execution when the FPU was disabled.
The cause was the use of a "#ifdef THINKOS_ENABLE_FPU" instead of "#if THINKOS_ENABLE_FPU" in the kernel.h.
The code was also improved a little bit by making sure the exceptions where disabled until the scheduler return.
Also the temporary register R2 was replaced by R6 in some parts of the scheduler to preserve R2 which is
used as parameter for the scheduler trace (dbg) functions.