Flash driver with no monitor direct API.
This is a test version. The monitor executes an auxiliary thread which
handles the flash operations and report back the result upon
completion.

This approach should reduce the complexity of the code.
There is no need for a dual API (user/kernel) for the flash.
All it's need is the user API.
This makes sense since when the monitor is loading a new firmware
as well as erasing the flash no application should be running.